start_watching() to autokindle/watchdog
