+++ autokindle/watchdog/__init__.py
@@ -1,0 +1,1 @@
+from autokindle.watchdog.utils import start_watching
